Remove useless `make()` and `unmake()` calls.
I assume they used to do something productive, but do not anymore.

This is not exposed as public API, however it is possible that a subclass
exists in someone's codebase that does something in these methods - but
it's not documented in any way, so I think getting rid of it is smart.

If anyone objects to this change, then perhaps we can revert this commit
and document the purpose of the functions so that it is known.