Don't allow Date.toString to be removed by DCE and fix PHP issues
Date.toString can be removed by DCE when targeting PHP and in
record-macros we typically do the equivalent to:

	class Main {
		static function main()
		{
			Sys.println(Std.string((Date.now():Dynamic)));
		}
	}

To fix the problem and make sure Date.toString is never removed (when
using this library), add @:keep through a initialization macro.

This however requires changing the test and travis files to use
record-macros as a library, instead of directly as a class path.

All test now pass PHP, and have been re-enabled.